Disable metadata endpoint
The metadata endpoint has a nasty bug causing HOPE jobs to fail. It's a tough
one to debug and I don't have time. Disable it so jobs pass.

This can be looked at again when hommod and yascript are running on the new
server and inside docker, making it easier to debug.